% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGraphFeatures.R
\name{compute_motif_counts}
\alias{compute_motif_counts}
\title{Count graph motifs (triangles, 3-stars, open 3-paths)}
\usage{
compute_motif_counts(adjacency)
}
\arguments{
\item{adjacency}{Adjacency matrix.}
}
\value{
Named list: \code{triangles}, \code{stars}, \code{paths}.
}
\description{
Count graph motifs (triangles, 3-stars, open 3-paths)
}
\keyword{internal}
