% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGraphFeatures.R
\name{compute_mst_diameter}
\alias{compute_mst_diameter}
\title{Compute MST diameter via double-DFS}
\usage{
compute_mst_diameter(mst_edges, mst_weights)
}
\arguments{
\item{mst_edges}{Integer matrix (\emph{(n-1)} \eqn{\times} 2) of MST edges.}

\item{mst_weights}{Numeric vector of corresponding edge weights.}
}
\value{
Numeric MST diameter (longest weighted path).
}
\description{
Compute MST diameter via double-DFS
}
\keyword{internal}
