% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGraphFeatures.R
\name{compute_transitivity}
\alias{compute_transitivity}
\title{Compute global transitivity (clustering coefficient)}
\usage{
compute_transitivity(adjacency)
}
\arguments{
\item{adjacency}{Logical or binary adjacency matrix.}
}
\value{
Numeric scalar in \eqn{[0, 1]}.
}
\description{
Compute global transitivity (clustering coefficient)
}
\keyword{internal}
