% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-graph.R
\name{.assert_igraph}
\alias{.assert_igraph}
\title{Assert that an object is an igraph}
\usage{
.assert_igraph(g, caller = "")
}
\arguments{
\item{g}{Object to check.}

\item{caller}{Character(1). Name of calling function for error messages.}
}
\description{
Assert that an object is an igraph
}
\keyword{internal}
